pyproject.toml
[project]
name = "agent-log-gif"
version = "0.6"
description = "Convert Claude Code session files to animated gifs and videos. Fork of https://github.com/simonw/claude-code-transcripts."
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Yoni Samlan" }
]
requires-python = ">=3.10"
dependencies = [
    "click",
    "click-default-group",
    "httpx",
    "Pillow>=10.0",
    "questionary",
]

[project.urls]
Homepage = "https://github.com/ysamlan/agent-log-gif"
Changelog = "https://github.com/ysamlan/agent-log-gif/releases"
Issues = "https://github.com/ysamlan/agent-log-gif/issues"
CI = "https://github.com/ysamlan/agent-log-gif/actions"

[project.scripts]
agent-log-gif = "agent_log_gif:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-httpx>=0.35.0",
    "ruff>=0.14.0",
]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]  # Line length handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
